changed the id type from int to long, to increase the db capacity

+++ TravelBuddyApi/Models/Message.cs
@@ -5,17 +5,17 @@ using System.Collections;
 
 public class Message
 {
-    public int MessageId { get; set; }
+    public long MessageId { get; set; }
 
     [Required]
     [MaxLength(800)]
     public string MessageContent { get; set; } = null!;
 
     /*FK -> Navigate to the Message owner*/
-    public int SenderId { get; set; }
+    public long SenderId { get; set; }
     public User Sender { get; set; } = null!;
 
-    public int ReceiverId { get; set; }
+    public long ReceiverId { get; set; }
     public User Receiver { get; set; } = null!;
 
     public DateTime SentAt { get; set; } = DateTime.UtcNow;
